Temporarily disable Sentry and PWA wrappers in next.config.mjs to isolate build error. Update Body component to accept font variables for improved font management. Wrap SkillsAndCases component in Suspense for better loading handling in ProjectsPage. Adjust service worker and source map for consistency.
