+++ components/pages/Body.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import localFont from 'next/font/local';
-import { Inter_Tight } from 'next/font/google';
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { Analytics } from '@vercel/analytics/react';
 import { Toaster } from '@/components/ui/sonner';
@@ -10,43 +8,18 @@ import Footer from '@/components/pages/Footer';
 import AmbientBackground from '@/components/pages/AmbientBackground';
 import CursorGlow from '@/components/pages/CursorGlow';
 
-const sourceSansPro = localFont({
-  src: [
-    {
-      path: '../../public/fonts/SourceSansPro-Regular.woff2',
-      weight: '400',
-      style: 'normal',
-    },
-    {
-      path: '../../public/fonts/SourceSansPro-Light.woff2',
-      weight: '300',
-      style: 'normal',
-    },
-    {
-      path: '../../public/fonts/SourceSansPro-Bold.woff2',
-      weight: '700',
-      style: 'normal',
-    },
-  ],
-  display: 'swap',
-  variable: '--font-source-sans',
-});
-
-// Map Inter Tight onto the legacy --font-oswald variable so existing classes continue to work
-const interTight = Inter_Tight({
-  subsets: ['latin'],
-  display: 'swap',
-  variable: '--font-oswald',
-});
-
 export default function Body({
   children,
+  fontVars,
 }: {
   children: React.ReactNode;
+  fontVars?: string;
 }) {
   return (
     <body
-      className={`dark ${sourceSansPro.variable} ${interTight.variable} flex flex-col md:flex-row min-h-screen dark:bg-gray-950 text-gray-200`}
+      className={`dark ${
+        fontVars ?? ''
+      } flex flex-col md:flex-row min-h-screen dark:bg-gray-950 text-gray-200`}
     >
       <AmbientBackground />
       <CursorGlow />
