+++ public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,a={};const c=(c,s)=>(c=new URL(c+".js",s).href,a[c]||new Promise(a=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=a,document.head.appendChild(e)}else e=c,importScripts(c),a()}).then(()=>{let e=a[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e}));self.define=(s,i)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(a[n])return;let t={};const f=e=>c(e,n),r={module:{uri:n},exports:t,require:f};a[n]=Promise.all(s.map(e=>r[e]||f(e))).then(e=>(i(...e),t))}}define(["./workbox-e9849328"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"f5b8a8d15124251bcdc9ed76c554511c"},{url:"/_next/dynamic-css-manifest.json",revision:"d751713988987e9331980363e24189ce"},{url:"/_next/static/A1AKv7MkyfcTczlPwbhmC/_buildManifest.js",revision:"c9387963ff3643dc7f21ccae61e42807"},{url:"/_next/static/A1AKv7MkyfcTczlPwbhmC/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/193-65d7a0dee3acb393.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/193-65d7a0dee3acb393.js.map",revision:"cecf80a5215aea75754229f9f6a575c2"},{url:"/_next/static/chunks/1e20724f-6e44c4ff6012fd36.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/1e20724f-6e44c4ff6012fd36.js.map",revision:"8df7366cea71a7573e3e1ea79dbfaf34"},{url:"/_next/static/chunks/229-eaa97941c51dfcc2.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/229-eaa97941c51dfcc2.js.map",revision:"c5126e5f4969cfbae2c93f8528d64498"},{url:"/_next/static/chunks/274-dd00b06345f86a2c.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/274-dd00b06345f86a2c.js.map",revision:"eaf6e1addf801bf2e1c8738982b50ed2"},{url:"/_next/static/chunks/745.78f5540901495013.js",revision:"78f5540901495013"},{url:"/_next/static/chunks/745.78f5540901495013.js.map",revision:"1e215b62c6c948654ab1ca09901d5fa2"},{url:"/_next/static/chunks/764-2df58a5eaaa81dc0.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/764-2df58a5eaaa81dc0.js.map",revision:"340337bf4182c84ff6ed99875e173739"},{url:"/_next/static/chunks/809-e9c7e0385b2e59e3.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/809-e9c7e0385b2e59e3.js.map",revision:"eec71d04c8752c9d93b1b030c1b561d2"},{url:"/_next/static/chunks/853-57cf30ea560454ce.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/853-57cf30ea560454ce.js.map",revision:"be09dcf146953a8abf4761ed336a99ce"},{url:"/_next/static/chunks/948-47ba07962f4a0973.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/948-47ba07962f4a0973.js.map",revision:"9eeae9c4ec500aa86be3ddab64957ac5"},{url:"/_next/static/chunks/app/_not-found/page-63d0cfaf7c476042.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/api/send/route-e028e7840015ce9f.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-02eed238478a6bfa.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/global-error-d9a7e13ab2de60b1.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/global-error-d9a7e13ab2de60b1.js.map",revision:"dee18c646ad1a785ec706953f424d8ce"},{url:"/_next/static/chunks/app/lab/page-28aeaa1c18cced30.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/layout-89491e67cf3a25e2.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/layout-89491e67cf3a25e2.js.map",revision:"f56355ee55a0d69b718ce4fe89d6efce"},{url:"/_next/static/chunks/app/not-found-3be503266099ab45.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/not-found-3be503266099ab45.js.map",revision:"9bc2dd275ce3e7600aa4fccb708fe4c0"},{url:"/_next/static/chunks/app/page-119e9e4837d37b51.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/page-119e9e4837d37b51.js.map",revision:"e6ef18dd190d637711967b38c37ecae1"},{url:"/_next/static/chunks/app/projects/page-39c8a1cd77d93ae2.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/app/projects/page-39c8a1cd77d93ae2.js.map",revision:"23ded8c9b2b008eb663182917ca8c5d0"},{url:"/_next/static/chunks/app/work/page-bc0f39ae3fc300fe.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/b9de3222-dd8c2feb484b6d73.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/b9de3222-dd8c2feb484b6d73.js.map",revision:"b243836172711eb5c68b7d900cfb2820"},{url:"/_next/static/chunks/framework-e550719c38f99844.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/framework-e550719c38f99844.js.map",revision:"f2d1653757f6dc99d4a950db61ec6eba"},{url:"/_next/static/chunks/main-22cb7741d0dfa1ba.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/main-22cb7741d0dfa1ba.js.map",revision:"0dfa35cbe184f6a77f57f85edebe85de"},{url:"/_next/static/chunks/main-app-2bf3f646e7900c3f.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/main-app-2bf3f646e7900c3f.js.map",revision:"e160d65ae1e48d50d8e270356586ba9e"},{url:"/_next/static/chunks/pages/_app-7a0d5e7491fa64c8.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/pages/_app-7a0d5e7491fa64c8.js.map",revision:"ea11d3c4b6c4f6cf4edc66665b8494d8"},{url:"/_next/static/chunks/pages/_error-318614c5f800d5f4.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/pages/_error-318614c5f800d5f4.js.map",revision:"c2f33ce7fabc6452762ba256213edd4a"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-4109c2e09a13c302.js",revision:"A1AKv7MkyfcTczlPwbhmC"},{url:"/_next/static/chunks/webpack-4109c2e09a13c302.js.map",revision:"ae1ca1205599b0d0e5545cfd1363f21e"},{url:"/_next/static/css/dcc3c6f531f53a8b.css",revision:"dcc3c6f531f53a8b"},{url:"/_next/static/css/dcc3c6f531f53a8b.css.map",revision:"d3f3d5e161829bd5956337b94e6f0f8e"},{url:"/_next/static/media/103fc5fac08dcb15-s.p.woff2",revision:"239e3c121b9f93f650289374acbf6288"},{url:"/_next/static/media/10735954b8957dce-s.woff2",revision:"d2ec0e63a71b8e0df83c59b76a4df343"},{url:"/_next/static/media/2a4438a5567e5337-s.p.woff2",revision:"2dcbd94d72a987eaab471ff35cae014d"},{url:"/_next/static/media/32ec8d161d520037-s.woff2",revision:"20e85716ef3eabc59aa5594aadeaabb9"},{url:"/_next/static/media/3cf1d7830273e71a-s.p.woff2",revision:"cd52ec283a11a58e4ddb1c1ca95b43aa"},{url:"/_next/static/media/69c8297e91a203ed-s.woff2",revision:"93e07d2ef36b2ab4db42ddde4176ed79"},{url:"/_next/static/media/84792440f5374347-s.woff2",revision:"b3ffb77daf1ddd24172ec70dea9af7d9"},{url:"/_next/static/media/af51cf31879cc10a-s.p.woff2",revision:"5df8a17352b50de2a86e59e84eb0e079"},{url:"/_next/static/media/c95a17d6aa1ccc69-s.woff2",revision:"cf9060e338211dea20a93490cbac6770"},{url:"/_next/static/media/ceda3e828136e1b8-s.woff2",revision:"0ac2f562e942403cf3fac6ba41831b88"},{url:"/fonts/SourceSansPro-Bold.woff2",revision:"5df8a17352b50de2a86e59e84eb0e079"},{url:"/fonts/SourceSansPro-Light.woff2",revision:"cd52ec283a11a58e4ddb1c1ca95b43aa"},{url:"/fonts/SourceSansPro-Regular.woff2",revision:"2dcbd94d72a987eaab471ff35cae014d"},{url:"/fonts/oswald.woff2",revision:"faed5f9f00128031d8bdc8be88b29ae2"},{url:"/icons/apple-icon.png",revision:"949583ba2230dbba22b8349473091f42"},{url:"/icons/favicon-96x96.png",revision:"0fd51540811fed283736b204a7c2a8f3"},{url:"/icons/favicon.ico",revision:"0a952045a7507d69770f25768172c842"},{url:"/icons/favicon.svg",revision:"15bc34ace4b8361c932c926cec2b674d"},{url:"/icons/site.webmanifest",revision:"9090e674d41e265d5f03c180fe5f6721"},{url:"/icons/web-app-manifest-192x192.png",revision:"e16c87c3911364e85803d69c66e97028"},{url:"/icons/web-app-manifest-512x512.png",revision:"c271e897682036284fff4b48fbfb5384"},{url:"/images/404.webp",revision:"a90e6bea419ca1227bc88ce7c9098caa"},{url:"/images/home/barbenheimer.webp",revision:"dd39f92be7864921d9070e77d5972872"},{url:"/images/home/mm-church.webp",revision:"73f5698dbf42963e89cdb48854e9a5b7"},{url:"/images/home/placeholder-image.webp",revision:"f3fd3133c71884f7a37b5a9506d2cafa"},{url:"/images/home/profile-image.jpg",revision:"b18ef6f0e5f842fb742f6b9cf6e0d557"},{url:"/images/home/profile-image.webp",revision:"ea53ce4be795c9c67fa87585cba89ac2"},{url:"/images/home/webdevhub.png",revision:"48ec5fa208104a420c10877eaa5cf666"},{url:"/images/home/webdevhub.webp",revision:"12cb21c69301a5a4b8b1aacda1a01acd"},{url:"/images/home/yahoo-dsp.webp",revision:"048641a57825768fa6475578026adf6d"},{url:"/images/projects/barbenheimer-zed-theme.webp",revision:"547746d4b39af742738f0728d6303aaa"},{url:"/images/projects/sync-flow.png",revision:"13d620af0a0105b72086c71dc28b4a1e"},{url:"/manifest.json",revision:"fa5b0d6ecc63e0261965d2fe4726daa9"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:c,state:s})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;const a=e.pathname;return!a.startsWith("/api/auth/")&&!!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")},new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>!(self.origin===e.origin),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
+if(!self.define){let e,a={};const s=(s,i)=>(s=new URL(s+".js",i).href,a[s]||new Promise(a=>{if("document"in self){const e=document.createElement("script");e.src=s,e.onload=a,document.head.appendChild(e)}else e=s,importScripts(s),a()}).then(()=>{let e=a[s];if(!e)throw new Error(`Module ${s} didn’t register its module`);return e}));self.define=(i,c)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(a[n])return;let t={};const r=e=>s(e,n),f={module:{uri:n},exports:t,require:r};a[n]=Promise.all(i.map(e=>f[e]||r(e))).then(e=>(c(...e),t))}}define(["./workbox-e9849328"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"5156cc32f08bdeaccdb981c9bc64ad31"},{url:"/_next/dynamic-css-manifest.json",revision:"d751713988987e9331980363e24189ce"},{url:"/_next/static/chunks/128-0aad1644ed84a09f.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/128-0aad1644ed84a09f.js.map",revision:"7c5df190bde1293b14d8bff1914efb56"},{url:"/_next/static/chunks/1e20724f-6e44c4ff6012fd36.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/1e20724f-6e44c4ff6012fd36.js.map",revision:"8df7366cea71a7573e3e1ea79dbfaf34"},{url:"/_next/static/chunks/337-78b4c419e059430f.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/337-78b4c419e059430f.js.map",revision:"d862b7f790b6862e425ef54ebbea4666"},{url:"/_next/static/chunks/672-86b5ec29faafe1a9.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/672-86b5ec29faafe1a9.js.map",revision:"38e733a4ced651fd3505955de8c9b0dc"},{url:"/_next/static/chunks/754-03b5d1e9c2e06436.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/754-03b5d1e9c2e06436.js.map",revision:"dd598baf1aa03b583f35dce8dc4932ac"},{url:"/_next/static/chunks/853-d059bd1447b205af.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/853-d059bd1447b205af.js.map",revision:"90bcd60a847d435b34a78f3cd798604a"},{url:"/_next/static/chunks/948-82eae773393e9506.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/948-82eae773393e9506.js.map",revision:"eded3beed65962b5ee440323f09b7344"},{url:"/_next/static/chunks/969-ea23cb45c2fccb4f.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/969-ea23cb45c2fccb4f.js.map",revision:"ba8fb3cd6dfc5faacb1664a7f03f1cf3"},{url:"/_next/static/chunks/app/_not-found/page-4b509f0a6f18d77c.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/api/send/route-14142c0ae3f2fc33.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-192d802b496f47b4.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/global-error-d9a7e13ab2de60b1.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/global-error-d9a7e13ab2de60b1.js.map",revision:"dee18c646ad1a785ec706953f424d8ce"},{url:"/_next/static/chunks/app/layout-d32d4b8a9390e1c2.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/layout-d32d4b8a9390e1c2.js.map",revision:"cdedcbbf07997b20c5c8d4ba84b6657a"},{url:"/_next/static/chunks/app/not-found-4dabb97c1b8aae5a.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/not-found-4dabb97c1b8aae5a.js.map",revision:"13704772cb779e8553f13ae306183f34"},{url:"/_next/static/chunks/app/page-a85303a1cdd43501.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/page-a85303a1cdd43501.js.map",revision:"4c6fc83ff62b01283b564ac0b4c0c5ff"},{url:"/_next/static/chunks/app/projects/layout-4d14208e2b0550cf.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/projects/page-31f96e319ecf702a.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/projects/page-31f96e319ecf702a.js.map",revision:"228c5452985e1837687aee6126414b76"},{url:"/_next/static/chunks/app/work/page-e50abc54670f01bc.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/app/work/page-e50abc54670f01bc.js.map",revision:"afb5dbc404fa217ed8cb5b8561ab835a"},{url:"/_next/static/chunks/b9de3222-dd8c2feb484b6d73.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/b9de3222-dd8c2feb484b6d73.js.map",revision:"b243836172711eb5c68b7d900cfb2820"},{url:"/_next/static/chunks/framework-e550719c38f99844.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/framework-e550719c38f99844.js.map",revision:"f2d1653757f6dc99d4a950db61ec6eba"},{url:"/_next/static/chunks/main-app-1a15bfb3c73dbb55.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/main-app-1a15bfb3c73dbb55.js.map",revision:"68ae60e7f0a4c4407d3bf403dd2849da"},{url:"/_next/static/chunks/main-e552c30f59b5ebae.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/main-e552c30f59b5ebae.js.map",revision:"6b04556522dd1f702dcf1703024b4f5e"},{url:"/_next/static/chunks/pages/_app-7a0d5e7491fa64c8.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/pages/_app-7a0d5e7491fa64c8.js.map",revision:"ea11d3c4b6c4f6cf4edc66665b8494d8"},{url:"/_next/static/chunks/pages/_error-318614c5f800d5f4.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/pages/_error-318614c5f800d5f4.js.map",revision:"c2f33ce7fabc6452762ba256213edd4a"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-9274c8793af5c7fd.js",revision:"xrj--L8RzFfwiqAjwX38t"},{url:"/_next/static/chunks/webpack-9274c8793af5c7fd.js.map",revision:"0675127d449e10dabe0f68de5a9d0650"},{url:"/_next/static/css/4541365b70099207.css",revision:"4541365b70099207"},{url:"/_next/static/css/4541365b70099207.css.map",revision:"882e513e586f41b2580eaf9e596d3a75"},{url:"/_next/static/media/103fc5fac08dcb15-s.p.woff2",revision:"239e3c121b9f93f650289374acbf6288"},{url:"/_next/static/media/10735954b8957dce-s.woff2",revision:"d2ec0e63a71b8e0df83c59b76a4df343"},{url:"/_next/static/media/2a4438a5567e5337-s.p.woff2",revision:"2dcbd94d72a987eaab471ff35cae014d"},{url:"/_next/static/media/32ec8d161d520037-s.woff2",revision:"20e85716ef3eabc59aa5594aadeaabb9"},{url:"/_next/static/media/3cf1d7830273e71a-s.p.woff2",revision:"cd52ec283a11a58e4ddb1c1ca95b43aa"},{url:"/_next/static/media/69c8297e91a203ed-s.woff2",revision:"93e07d2ef36b2ab4db42ddde4176ed79"},{url:"/_next/static/media/84792440f5374347-s.woff2",revision:"b3ffb77daf1ddd24172ec70dea9af7d9"},{url:"/_next/static/media/af51cf31879cc10a-s.p.woff2",revision:"5df8a17352b50de2a86e59e84eb0e079"},{url:"/_next/static/media/c95a17d6aa1ccc69-s.woff2",revision:"cf9060e338211dea20a93490cbac6770"},{url:"/_next/static/media/ceda3e828136e1b8-s.woff2",revision:"0ac2f562e942403cf3fac6ba41831b88"},{url:"/_next/static/xrj--L8RzFfwiqAjwX38t/_buildManifest.js",revision:"2f1884e02d715201b390b11e80c38e09"},{url:"/_next/static/xrj--L8RzFfwiqAjwX38t/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/fonts/SourceSansPro-Bold.woff2",revision:"5df8a17352b50de2a86e59e84eb0e079"},{url:"/fonts/SourceSansPro-Light.woff2",revision:"cd52ec283a11a58e4ddb1c1ca95b43aa"},{url:"/fonts/SourceSansPro-Regular.woff2",revision:"2dcbd94d72a987eaab471ff35cae014d"},{url:"/fonts/oswald.woff2",revision:"faed5f9f00128031d8bdc8be88b29ae2"},{url:"/icons/apple-icon.png",revision:"949583ba2230dbba22b8349473091f42"},{url:"/icons/favicon-96x96.png",revision:"0fd51540811fed283736b204a7c2a8f3"},{url:"/icons/favicon.ico",revision:"0a952045a7507d69770f25768172c842"},{url:"/icons/favicon.svg",revision:"15bc34ace4b8361c932c926cec2b674d"},{url:"/icons/site.webmanifest",revision:"9090e674d41e265d5f03c180fe5f6721"},{url:"/icons/web-app-manifest-192x192.png",revision:"e16c87c3911364e85803d69c66e97028"},{url:"/icons/web-app-manifest-512x512.png",revision:"c271e897682036284fff4b48fbfb5384"},{url:"/images/404.webp",revision:"a90e6bea419ca1227bc88ce7c9098caa"},{url:"/images/home/barbenheimer.webp",revision:"dd39f92be7864921d9070e77d5972872"},{url:"/images/home/mm-church.webp",revision:"73f5698dbf42963e89cdb48854e9a5b7"},{url:"/images/home/placeholder-image.webp",revision:"f3fd3133c71884f7a37b5a9506d2cafa"},{url:"/images/home/profile-image.jpg",revision:"b18ef6f0e5f842fb742f6b9cf6e0d557"},{url:"/images/home/profile-image.webp",revision:"ea53ce4be795c9c67fa87585cba89ac2"},{url:"/images/home/webdevhub.png",revision:"48ec5fa208104a420c10877eaa5cf666"},{url:"/images/home/webdevhub.webp",revision:"12cb21c69301a5a4b8b1aacda1a01acd"},{url:"/images/home/yahoo-dsp.webp",revision:"048641a57825768fa6475578026adf6d"},{url:"/images/projects/aurora-borealis.png",revision:"e8abc7ed0a8ed50f656f613951f198c5"},{url:"/images/projects/barbenheimer-zed-theme.webp",revision:"547746d4b39af742738f0728d6303aaa"},{url:"/images/projects/christmas-countdown.png",revision:"1dfa37fe2d448645c74b74736ef685aa"},{url:"/images/projects/croo.png",revision:"2d3f0d635ff9a9f8912fca97a1510fe5"},{url:"/images/projects/eat-bulaga-mobile.png",revision:"e868952f644032c4327219224cc27290"},{url:"/images/projects/ember-upgrade-guide.png",revision:"776bfb98f2dee5ec2185a59ae3c372cc"},{url:"/images/projects/fintq-lendr.png",revision:"522dd3565aa20b4b1b1009cf2970ac1a"},{url:"/images/projects/pinoy-hoops.webp",revision:"91d39b69e17fb552c1312fc1c3fa872a"},{url:"/images/projects/solar-eclipse.png",revision:"c1cec0f5278d9d65ac26897b2a5f8635"},{url:"/images/projects/sync-flow.png",revision:"13d620af0a0105b72086c71dc28b4a1e"},{url:"/manifest.json",revision:"fa5b0d6ecc63e0261965d2fe4726daa9"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:s,state:i})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;const a=e.pathname;return!a.startsWith("/api/auth/")&&!!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")},new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>!(self.origin===e.origin),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
 //# sourceMappingURL=sw.js.map
